collection in collection invoice list in collection screen added

+++ app/src/main/java/com/example/myservice/data/model/InvoiceCollectionReq.kt
@@ -7,7 +7,7 @@ data class InvoiceCollectionReq(
     val partyId: Int,
 
     @SerializedName("collection_amount")
-    val collectionAmount: Double,
+    val collectionAmount: Int,
 
     @SerializedName("transaction_date")
     val transactionDate: String,
